Adding argument parser, update test
argument parser added to run.py to select the way to display datas (csv or database)
Update test file to match lasts modifications

+++ run.py
@@ -4,21 +4,39 @@ from apps.schools.models import School
 from apps.cities.controller import parse_csv_cities_to_database
 from apps.schools.controller import parse_csv_schools_to_database
 from database import db
+import argparse
 
 
 if __name__ == '__main__':
-    db.close()
-    db.connect()
-    q = City.delete()
-    q.execute()
-    q = School.delete()
-    q.execute()
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "action",
+        help="Choose an action to execute",
+        nargs="?",
+        choices=[
+            "from_csv",
+            "from_database"
+        ],
+    )
+    args = parser.parse_args()
 
-    parse_csv_cities_to_database()
+    if args.action == "from_csv":
+        merged_dataframe = merge_cities_and_schools()
+        show_graphe_merged_cities_and_schools(merged_dataframe)
 
-    merged_dataframe = merge_cities_and_schools()
+    if args.action == "from_database":
+        db.close()
+        db.connect()
+        q = City.delete()
+        q.execute()
+        q = School.delete()
+        q.execute()
 
-    parse_csv_schools_to_database(merged_dataframe)
+        parse_csv_cities_to_database()
 
-    show_graphe_merged_cities_and_schools(merged_dataframe)
-    db.close()
+        merged_dataframe = merge_cities_and_schools()
+
+        parse_csv_schools_to_database(merged_dataframe)
+
+        show_graphe_merged_cities_and_schools(merged_dataframe)
+        db.close()
